site/agents/requirements.txt
anthropic>=0.40.0
requests>=2.31.0
aiohttp>=3.9.0
scikit-learn>=1.3.0
python-frontmatter>=1.1.0
google-auth>=2.23.0
google-auth-httplib2>=0.1.1
python-dotenv>=1.0.0
